+++ src/main/groovy/nebula/plugin/dependencylock/tasks/GenerateLockTask.groovy
@@ -28,7 +28,7 @@ class GenerateLockTask extends AbstractLockTask {
     Set<String> configurationNames
     File dependenciesLock
     Map overrides
-    Boolean includeTransitives
+    boolean includeTransitives = false
 
     @TaskAction
     void lock() {
@@ -37,28 +37,36 @@ class GenerateLockTask extends AbstractLockTask {
     }
 
     private readDependenciesFromConfigurations() {
-        def deps = [:].withDefault { [via: [] as Set, childrenVisited: false] }
+        def deps = [:].withDefault { [transitive: [] as Set, firstLevelTransitive: [] as Set, childrenVisited: false] }
         def confs = getConfigurationNames().collect { project.configurations.getByName(it) }
 
+        def peers = project.rootProject.allprojects.collect { new LockKey(group: it.group, artifact: it.name) }
+
         confs.each { Configuration configuration ->
-            def peers = configuration.allDependencies.withType(ProjectDependency).collect { new LockKey(group: it.group, artifact: it.name) }
             configuration.allDependencies.withType(ExternalDependency).each { Dependency dependency ->
                 def key = new LockKey(group: dependency.group, artifact: dependency.name)
-                deps[key.toString()].requested = dependency.version
+                deps[key].requested = dependency.version
             }
             configuration.resolvedConfiguration.firstLevelModuleDependencies.each { ResolvedDependency resolved ->
                 def key = new LockKey(group: resolved.moduleGroup, artifact: resolved.moduleName)
                 if (!peers.contains(key)) {
-                    deps[key.toString()].locked = resolved.moduleVersion
+                    deps[key].locked = resolved.moduleVersion
+                } else {
+                    deps[key].project = true
+                    if (!getIncludeTransitives()) {
+                        handleSiblingTransitives(resolved, deps, peers)
+                    }
                 }
                 if (getIncludeTransitives()) {
-                    deps[key.toString()].childrenVisited = true
-                    resolved.children.each { handleTransitive(it, deps, key.toString()) }
+                    deps[key].childrenVisited = true
+                    resolved.children.each { handleTransitive(it, deps, peers, key) }
                 }
             }
         }
 
-        getOverrides().each { String key, String overrideVersion ->
+        getOverrides().each { String k, String overrideVersion ->
+            def tokens = k.tokenize(':')
+            LockKey key = new LockKey(group: tokens[0], artifact: tokens[1] )
             if (deps.containsKey(key)) {
                 deps[key].viaOverride = overrideVersion
             }
@@ -67,19 +75,41 @@ class GenerateLockTask extends AbstractLockTask {
         return deps
     }
 
-    private static handleTransitive(ResolvedDependency transitive, Map deps, String parent) {
-        def key = new LockKey(group: transitive.moduleGroup, artifact: transitive.moduleName).toString()
+    private static handleSiblingTransitives(ResolvedDependency sibling, Map deps, List peers) {
+        def parent = new LockKey(group: sibling.moduleGroup, artifact: sibling.moduleName)
+        sibling.children.each { ResolvedDependency dependency ->
+            def key = new LockKey(group: dependency.moduleGroup, artifact: dependency.moduleName)
+            deps[key].firstLevelTransitive << parent
+            if (peers.contains(key) && !deps[key].childrenVisited) {
+                if (dependency.children.size() > 0) {
+                    deps[key].childrenVisited = true
+                    handleSiblingTransitives(dependency, deps, peers)
+                }
+            } else {
+                deps[key].locked = dependency.moduleVersion
+            }
+        }
+    }
+
+    private static handleTransitive(ResolvedDependency transitive, Map deps, List peers, LockKey parent) {
+        def key = new LockKey(group: transitive.moduleGroup, artifact: transitive.moduleName)
 
         if (!deps[key].childrenVisited) {
-            deps[key].locked = transitive.moduleVersion
-            deps[key].transitive = true
-            transitive.children.each { handleTransitive(it, deps, key) }
+            if (!peers.contains(key)) {
+                deps[key].locked = transitive.moduleVersion
+            } else {
+                deps[key].project = true
+            }
+            if (transitive.children.size() > 0) {
+                deps[key].childrenVisited = true
+            }
+            transitive.children.each { handleTransitive(it, deps, peers, key) }
         }
-        deps[key].via << parent
+        deps[key].transitive << parent
     }
 
     private void writeLock(deps) {
-        def strings = deps.collect { String k, Map v -> stringifyLock(k, v) }
+        def strings = deps.collect { LockKey k, Map v -> stringifyLock(k, v) }
         strings = strings.sort()
         project.buildDir.mkdirs()
         getDependenciesLock().withPrintWriter { out ->
@@ -89,20 +119,26 @@ class GenerateLockTask extends AbstractLockTask {
         }
     }
 
-    private static String stringifyLock(String key, Map lock) {
-        def lockLine = new StringBuilder("  \"${key}\": { \"locked\": \"${lock.locked}\"")
+    private static String stringifyLock(LockKey key, Map lock) {
+        def lockLine = new StringBuilder("  \"${key}\": { ")
+        if (lock.locked) {
+            lockLine << "\"locked\": \"${lock.locked}\""
+        } else {
+            lockLine << '"project": true'
+        }
         if (lock.requested) {
             lockLine << ", \"requested\": \"${lock.requested}\""
-        }
-        if (lock.transitive) {
-            lockLine << ", \"transitive\": true"
         }
         if (lock.viaOverride) {
             lockLine << ", \"viaOverride\": \"${lock.viaOverride}\""
         }
-        if (lock.via) {
-            def transitiveFrom = lock.via.sort().collect { "\"${it}\""}.join(', ')
-            lockLine << ", \"via\": [ ${transitiveFrom} ]"
+        if (lock.transitive) {
+            def transitiveFrom = lock.transitive.sort().collect { "\"${it}\""}.join(', ')
+            lockLine << ", \"transitive\": [ ${transitiveFrom} ]"
+        }
+        if (lock.firstLevelTransitive) {
+            def transitiveFrom = lock.firstLevelTransitive.sort().collect { "\"${it}\""}.join(', ')
+            lockLine << ", \"firstLevelTransitive\": [ ${transitiveFrom} ]"
         }
         lockLine << ' }'