+++ src/test/groovy/nebula/plugin/dependencylock/DependencyLockPluginSpec.groovy
@@ -84,6 +84,25 @@ class DependencyLockPluginSpec extends ProjectSpec {
         foo.moduleVersion == '2.0.1'
     }
 
+    def 'command line override of a dependency works with no other dependency lock features'() {
+        project.apply plugin: 'java'
+        project.repositories { maven { url Fixture.repo } }
+        project.dependencies {
+            compile 'test.example:foo:1.+'
+        }
+
+        project.ext.set('dependencyLock.override', 'test.example:foo:1.0.0')
+
+        when:
+        project.apply plugin: pluginName
+        triggerTaskGraphWhenReady()
+
+        then:
+        def resolved = project.configurations.compile.resolvedConfiguration
+        def foo = resolved.firstLevelModuleDependencies.find { it.moduleName == 'foo' }
+        foo.moduleVersion == '1.0.0'
+    }
+
     def 'command line override of a dependency'() {
         stockTestSetup()
 
@@ -97,6 +116,33 @@ class DependencyLockPluginSpec extends ProjectSpec {
         then:
         def foo = resolved.firstLevelModuleDependencies.find { it.moduleName == 'foo' }
         foo.moduleVersion == '2.0.1'
+    }
+
+    def 'command line override of a dependency with forces in place'() {
+        stockTestSetup()
+
+        project.dependencies {
+            compile 'test.example:baz:1.+'
+        }
+
+        project.configurations.all {
+            resolutionStrategy {
+                force 'test.example:baz:1.0.0'
+            }
+        }
+
+        project.ext.set('dependencyLock.override', 'test.example:foo:2.0.1')
+
+        when:
+        project.apply plugin: pluginName
+        triggerTaskGraphWhenReady()
+        def resolved = project.configurations.compile.resolvedConfiguration
+
+        then:
+        def foo = resolved.firstLevelModuleDependencies.find { it.moduleName == 'foo' }
+        foo.moduleVersion == '2.0.1'
+        def baz = resolved.firstLevelModuleDependencies.find { it.moduleName == 'baz' }
+        baz.moduleVersion == '1.0.0'
     }
 
     def 'command line overrides of multiple dependencies'() {
@@ -216,6 +262,56 @@ class DependencyLockPluginSpec extends ProjectSpec {
         bar2 == null
         def baz2 = resolved2.firstLevelModuleDependencies.find { it.moduleName == 'baz' }
         baz2.moduleVersion == '1.0.0'
+    }
+
+    def 'siblings need to lock each others dependencies'() {
+        def sub1Folder = new File(projectDir, 'sub1')
+        sub1Folder.mkdir()
+        def sub1 = ProjectBuilder.builder().withName('sub1').withProjectDir(sub1Folder).withParent(project).build()
+        def sub1DependenciesLock = new File(sub1Folder, 'dependencies.lock')
+        sub1DependenciesLock << '''\
+            {
+              "test.example:baz": { "locked": "1.0.0", "requested": "1.+" }
+            }
+        '''.stripIndent()
+
+        def sub2Folder = new File(projectDir, 'sub2')
+        sub2Folder.mkdir()
+        def sub2 = ProjectBuilder.builder().withName('sub2').withProjectDir(sub2Folder).withParent(project).build()
+        def sub2DependenciesLock = new File(sub2Folder, 'dependencies.lock')
+        sub2DependenciesLock << '''\
+            {
+              "test.example:baz": { "locked": "1.0.0", "firstLevelTransitive": [ "test.nebula:sub1" ] },
+              "test.example:foo": { "locked": "2.0.0", "requested": "2.+" },
+              "test.nebula:sub1": { "project": true, "firstLevelTransitive": [ "test.nebula:sub2" ] }
+            }
+        '''.stripIndent()
+
+        project.subprojects {
+            apply plugin: 'java'
+            group = 'test.nebula'
+            repositories { maven { url Fixture.repo } }
+        }
+
+        sub1.dependencies {
+            compile 'test.example:baz:1.+'
+        }
+
+        sub2.dependencies {
+            compile project.project(':sub1')
+            compile 'test.example:foo:2.+'
+        }
+
+        when:
+        project.subprojects {
+            apply plugin: pluginName
+        }
+        triggerTaskGraphWhenReady()
+
+        then:
+        def resolved2 = sub2.configurations.compile.resolvedConfiguration
+        def baz = resolved2.resolvedArtifacts.find { it.name == 'baz' }
+        baz.moduleVersion.id.version == '1.0.0'
     }
 
     private List multiProjectSetup() {