+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+1.9.9 / 2014-06-03
+==================
+
+* Allow command line overrides to work if plugin is applied, not just when creating a lock or in the presence of a lock.
+* Bugfix: lock first level transitives of any project(':<project>') depended on
+* Bugfix: properly respect resolutionStrategy.force 'group:artifact:version' present in build.gradle
+* Add commitLock task
+
+1.9.8 / 2014-05-02
+==================
+
+* Bug fix: properly skip interproject dependencies in a multi-project build when including transitives in the lock
+
+1.9.7 / 2014-04-30
+==================
+
+* Fix handling of circular references in transitive dependencies
+
 1.9.6 / 2014-04-28
 ==================